move authenticator.use into config
