+++ remix/app/routes/auth.google.tsx
@@ -1,5 +1,5 @@
 import type { ActionArgs } from "@remix-run/node";
-import { authenticator } from "~/services/auth.server";
+import { authenticator } from "~/config/auth.server";
 
 import { SocialsProvider } from "remix-auth-socials";
