+++ remix/app/services/auth.server.ts
@@ -1,13 +1,7 @@
-import { sessionStorage } from "./session.server";
 import type { UserProfile } from "~/types";
 import { db } from "~/config/db.server";
 
-import { Authenticator } from "remix-auth";
-import { GoogleStrategy, SocialsProvider } from "remix-auth-socials";
-
-export const authenticator = new Authenticator(sessionStorage);
-
-async function handleSocialAuthCallBack({ profile }: any) {
+export async function handleSocialAuthCallBack({ profile }: any) {
   const userProfile: UserProfile = profile?._json;
 
   if (userProfile) {
@@ -32,15 +26,3 @@ async function handleSocialAuthCallBack({ profile }: any) {
   }
   return profile;
 }
-
-authenticator.use(
-  new GoogleStrategy(
-    {
-      clientID: process.env.GOOGLE_CLIENT_ID || "",
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET || "",
-      scope: ["profile", "email", "openid"],
-      callbackURL: `http://localhost:3400/auth/${SocialsProvider.GOOGLE}/callback`,
-    },
-    handleSocialAuthCallBack
-  )
-);
